Muevo algunas funciones a otros archivos para evitar archivos grandes
refs:#87
