+++ tests/factories/utils.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+
+def jsonschema_str(string):
+    return repr(string)
