--- tests/results/invalid_catalog_publisher_type.json
@@ -1,33 +1,0 @@
-{
-    "status": "ERROR",
-    "error": {
-        "catalog": {
-            "status": "ERROR",
-            "errors": [
-                {
-                    "instance": [
-                        "Ministerio de Modernización",
-                        "datos@modernizacion.gob.ar"
-                    ],
-                    "validator": "type",
-                    "path": [
-                        "publisher"
-                    ],
-                    "message": "[u'Ministerio de Modernizaci\\xf3n', u'datos@modernizacion.gob.ar'] is not of type u'object'",
-                    "error_code": 2,
-                    "validator_value": "object"
-                }
-            ],
-            "title": "Datos Argentina"
-        },
-        "dataset": [
-            {
-                "status": "OK",
-                "identifier": "99db6631-d1c9-470b-a73e-c62daa32c420",
-                "list_index": 0,
-                "errors": [],
-                "title": "Sistema de contrataciones electrónicas"
-            }
-        ]
-    }
-}
