--- tests/results/invalid_publisher_mbox_format.json
@@ -1,31 +1,0 @@
-{
-    "status": "ERROR",
-    "error": {
-        "catalog": {
-            "status": "ERROR",
-            "errors": [
-                {
-                    "instance": "datosATmodernizacion.gob.ar",
-                    "validator": "format",
-                    "path": [
-                        "publisher",
-                        "mbox"
-                    ],
-                    "message": "u'datosATmodernizacion.gob.ar' is not a u'email'",
-                    "error_code": 2,
-                    "validator_value": "email"
-                }
-            ],
-            "title": "Datos Argentina"
-        },
-        "dataset": [
-            {
-                "status": "OK",
-                "identifier": "99db6631-d1c9-470b-a73e-c62daa32c420",
-                "list_index": 0,
-                "errors": [],
-                "title": "Sistema de contrataciones electrónicas"
-            }
-        ]
-    }
-}
